shifted data model related methods from Series to PieSeries
