+++ org.eclipse.swtchart/src/org/eclipse/swtchart/IPieSeries.java
@@ -1,5 +1,12 @@
 package org.eclipse.swtchart;
 
+import org.eclipse.swt.graphics.Color;
 
 public interface IPieSeries {
+	
+	public String[] getLabelSeries();
+	
+	public double[] getValueSeries();
+	
+	public Color[] getColors();
 }
