+++ org.eclipse.swtchart/src/org/eclipse/swtchart/internal/series/Series.java
@@ -29,14 +29,11 @@ import org.eclipse.swtchart.ISeries;
 import org.eclipse.swtchart.ISeriesLabel;
 import org.eclipse.swtchart.Range;
 import org.eclipse.swtchart.internal.axis.Axis;
-import org.eclipse.swtchart.internal.compress.CompressBarSeries;
-import org.eclipse.swtchart.internal.compress.CompressPieSeries;
 import org.eclipse.swtchart.internal.compress.ICompress;
 import org.eclipse.swtchart.model.CartesianSeriesModel;
 import org.eclipse.swtchart.model.DateArraySeriesModel;
 import org.eclipse.swtchart.model.DoubleArraySeriesModel;
 import org.eclipse.swtchart.model.IndexedSeriesModel;
-import org.eclipse.swtchart.model.StringArraySeriesModel;
 
 /**
  * Series.
@@ -76,7 +73,6 @@ abstract public class Series<T> implements ISeries<T> {
 	/** the list of dispose listeners */
 	private List<IDisposeListener> listeners;
 	private CartesianSeriesModel<T> model;
-	private StringArraySeriesModel stringArraySeriesModel;
 
 	/**
 	 * Constructor.
@@ -142,11 +138,6 @@ abstract public class Series<T> implements ISeries<T> {
 
 		return model;
 	}
-
-	public StringArraySeriesModel getStringArraySeriesModel() {
-		
-		return stringArraySeriesModel;
-	}
 	
 	@Override
 	public void setDataModel(CartesianSeriesModel<T> model) {
@@ -171,19 +162,6 @@ abstract public class Series<T> implements ISeries<T> {
 			stackEnabled = false;
 		}
 	}
-
-	public void setStringArrayModel(StringArraySeriesModel data) {
-		
-		this.stringArraySeriesModel = data;
-		setCompressor();
-		if(compressor instanceof CompressPieSeries) {
-			((CompressPieSeries)compressor).setLabelSeries(getLabelSeries());
-			((CompressPieSeries)compressor).setValueSeries(getValueSeries());
-		}
-		else if(compressor instanceof CompressBarSeries) {
-			//code for category series
-		}
-	}
 	
 	@Override
 	public void enableStack(boolean enabled) {
@@ -197,20 +175,6 @@ abstract public class Series<T> implements ISeries<T> {
 		}
 		stackEnabled = enabled;
 		((SeriesSet)chart.getSeriesSet()).updateStackAndRiserData();
-	}
-
-	public void setSeries(String[] labels, double[] values) {
-		
-		if(labels == null || values == null ) {
-			SWT.error(SWT.ERROR_NULL_ARGUMENT);
-			return; // to suppress warning...
-		}
-		String[] ids = new String[labels.length];
-		System.arraycopy(labels, 0, ids, 0, labels.length);
-		double[] val = new double[values.length];
-		System.arraycopy(values, 0, val, 0, values.length);
-		StringArraySeriesModel data = new StringArraySeriesModel(ids,val);
-		setStringArrayModel(data);
 	}
 	
 	@SuppressWarnings("unchecked")
@@ -262,22 +226,7 @@ abstract public class Series<T> implements ISeries<T> {
 		return StreamSupport.stream(dataModel.spliterator(), false).filter(t -> dataModel.getX(t) != null).mapToDouble(value -> dataModel.getX(value).doubleValue()).toArray();
 	}
 
-	public String[] getLabelSeries() {
-		StringArraySeriesModel stringArraySeriesModel = getStringArraySeriesModel();
-		String[] labels = stringArraySeriesModel.getLabels();
-		String[] ids = new String[labels.length];
-		System.arraycopy(labels, 0, ids, 0, labels.length);
-		return ids;
-	}
-	
-	public double[] getValueSeries() {
-		StringArraySeriesModel stringArraySeriesModel = getStringArraySeriesModel();
-		double[] values = stringArraySeriesModel.getValues();
-		double[] val = new double[values.length];
-		System.arraycopy(values, 0, val, 0, values.length);
-		return val;
-	}
-	
+
 	@SuppressWarnings("unchecked")
 	@Override
 	public void setYSeries(double[] series) {
